rename package vote for dayvoting
