+++ app/src/main/java/cz/devconf2017/dayvoting/TalkBusiness.java
@@ -1,4 +1,4 @@
-package cz.devconf2017.vote;
+package cz.devconf2017.dayvoting;
 
 import android.support.annotation.Nullable;
